fixing stop_gate()
Switching off both relays at the same time can cause a back current flow
that can break the direction relay and/or pop off the fuse.
To solve it, the motor is first stopped maintaining the same direction so only
one relay is changed, and after a comulative delay of at least `def_dir_delay`
the direction is changed.